Remove `AUTH_TOKEN`
This parameter is not used any more, and has caused a lot of confusion.

In the distant past, the dashboard was used at Mapzen to track progress
of builds. As part of this, a Redis instance stored how many documents
were seen in the _previous_ build (Mapzen had a dedicated build
cluster), and used that to compare against.

The auth token was use to access this Redis instance.

Because that workflow was heavily integrated into Mapzen's
infrastructure, and isn't really relevant to how we generally suggest
setting up Pelias these days, it's best to remove it all.